ENH: unifying panel, stiffener, stiffpanelbay methods
