+++ compmech/panel/__init__.py
@@ -5,12 +5,8 @@ Semi-analytical Models for Panels (:mod:`compmech.panel`)
 
 .. currentmodule:: compmech.panel
 
-.. automodule:: compmech.panel.kpanel
-    :members:
-
-.. automodule:: compmech.panel.cpanel
+.. automodule:: compmech.panel.panel
     :members:
 
 """
-import kpanel
-import cpanel
+from panel import Panel
