+++ compmech/panel/models/cpanel_clt_donnell_bardell_field.pyx
