+++ compmech/stiffener/__init__.py
@@ -1,0 +1,18 @@
+r"""
+==============================================================================
+Stiffeners (:mod:`compmech.stiffener`)
+==============================================================================
+
+.. currentmodule:: compmech.stiffener
+
+This module has all stiffener's formulations used along other classes.
+
+.. autoclass:: BladeStiff1D
+    :members:
+
+.. autoclass:: BladeStiff2D
+    :members:
+
+
+"""
+from stiffener import BladeStiff1D, BladeStiff2D
