+++ compmech/stiffener/modelDB.py
@@ -1,19 +1,18 @@
 r"""
 Used to configure the main parameters for each implemented model.
 
-.. currentmodule:: compmech.panel.modelDB
+.. currentmodule:: compmech.stiffener.modelDB
 
 
 """
 from models import *
 
 db = {
-    'plate_clt_donnell_bardell': {
+    'bladestiff2d_clt_donnell_bardell': {
                     'linear static': True,
                     'linear buckling': True,
                     'non-linear static': False,
-                    'field': plate_clt_donnell_bardell_field,
-                    'matrices': plate_clt_donnell_bardell,
+                    'matrices': bladestiff2d_clt_donnell_bardell,
                     'dofs': 3,
                     'e_num': 6,
                     'num': 3,
