+++ compmech/stiffener/models/__init__.py
