+++ compmech/stiffener/setup.py
@@ -5,8 +5,8 @@ import sys
 
 def configuration(parent_package='',top_path=None):
     from numpy.distutils.misc_util import Configuration
-    config = Configuration('stiffpanelbay2d',parent_package,top_path)
-    config.add_subpackage('clpt')
+    config = Configuration('stiffener',parent_package,top_path)
+    config.add_subpackage('models')
     config.make_config_py()
     return config
