--- compmech/stiffpanel1d/fsdt/fsdt_commons_bc1.pxd
@@ -1,3 +1,0 @@
-cdef void cfuvw(double *c, int m1, int n1, double a, double b, double *xs,
-        double *ys, int size, double *us, double *vs, double *ws,
-        double *phixs, double *phiys) nogil
