--- compmech/stiffpanelbay2d/clpt/setup.bat
@@ -1,1 +1,0 @@
-python setup.py build_ext -IC:\repos\compmech\C\include -LC:\repos\compmech\C\lib --force --inplace clean
